Add module for confluent schema registry
