+++ dd-java-agent/instrumentation/confluent-schema-registry/confluent-schema-registry-4.1/src/main/java/datadog/trace/instrumentation/confluentschemaregistry/KafkaDeserializerInstrumentation.java
@@ -6,15 +6,11 @@ import static net.bytebuddy.matcher.ElementMatchers.isPublic;
 import static net.bytebuddy.matcher.ElementMatchers.takesArgument;
 import static net.bytebuddy.matcher.ElementMatchers.takesArguments;
 
-import com.google.auto.service.AutoService;
 import datadog.trace.agent.tooling.Instrumenter;
 import datadog.trace.agent.tooling.Instrumenter.MethodTransformer;
-import datadog.trace.agent.tooling.InstrumenterModule;
 import datadog.trace.bootstrap.InstrumentationContext;
 import datadog.trace.bootstrap.instrumentation.api.AgentTracer;
 import datadog.trace.instrumentation.kafka_common.ClusterIdHolder;
-import java.util.HashMap;
-import java.util.Map;
 import net.bytebuddy.asm.Advice;
 import org.apache.kafka.common.serialization.Deserializer;
 
@@ -22,13 +18,8 @@ import org.apache.kafka.common.serialization.Deserializer;
  * Instruments Confluent Schema Registry deserializers (Avro, Protobuf, and JSON) to capture
  * deserialization operations.
  */
-@AutoService(InstrumenterModule.class)
-public class KafkaDeserializerInstrumentation extends InstrumenterModule.Tracing
+public class KafkaDeserializerInstrumentation
     implements Instrumenter.ForKnownTypes, Instrumenter.HasMethodAdvice {
-
-  public KafkaDeserializerInstrumentation() {
-    super("confluent-schema-registry", "kafka");
-  }
 
   @Override
   public String[] knownMatchingTypes() {
@@ -37,21 +28,6 @@ public class KafkaDeserializerInstrumentation extends InstrumenterModule.Tracing
       "io.confluent.kafka.serializers.json.KafkaJsonSchemaDeserializer",
       "io.confluent.kafka.serializers.protobuf.KafkaProtobufDeserializer"
     };
-  }
-
-  @Override
-  public String[] helperClassNames() {
-    return new String[] {
-      "datadog.trace.instrumentation.kafka_common.ClusterIdHolder",
-      packageName + ".SchemaIdExtractor"
-    };
-  }
-
-  @Override
-  public Map<String, String> contextStore() {
-    Map<String, String> contextStores = new HashMap<>();
-    contextStores.put("org.apache.kafka.common.serialization.Deserializer", "java.lang.Boolean");
-    return contextStores;
   }
 
   @Override
